Update permissions.md
Under Example 1:
I removed an extraneous space in the list of permissions.
Also, I believe that the last bullet, "Effective permissions for JSmith" was inaccurate so I added the "delete" permission to the end.   Otherwise, the last bullet and the final note in that section seemed contradictory.  Please review, and double-check, as I want to make sure I am interpreting it correctly.