Update create-tenants.md
The UI Management Form section did not address the option to Create a new FQDN for a tenant.  I added that option on the page and tried to create a link to the Knowledge Base document that describes how to set that up using a proxy.  Please review and verify that the link is set up correctly.